ip2gid: Move msg_length_check() to a separate file
Move msg_length_check() to a separate file msg_check.c.
