+++ CMakeLists.txt
@@ -27,7 +27,7 @@ endif()
 set(CMAKE_C_COMPILER_NAMES gcc clang)
 include_directories("include")
 add_executable(ip2gid
-	src/main.c src/log.c)
+	src/main.c src/log.c src/msg_check.c)
 target_compile_options(ip2gid PRIVATE -Wall)
 
 set(SEARCHPATHS
