+++ src/main.c
@@ -91,85 +91,6 @@ static void clear_timeout_cell_req(void)
 		   cells_used, DEFAULT_PENDING_REQUESTS);
 }
 
-static int msg_length_check(struct ip2gid_obj *obj, uint32_t max_len)
-{
-	struct ip2gid_tlv_hdr *tlv;
-	struct ip2gid_hdr *hdr;
-	uint16_t num_tlvs;
-
-	if (max_len <= sizeof(struct ip2gid_hdr)) {
-		ip2gid_log(IP2GID_LOG_INFO,
-			   "Msg length too short\n");
-		return -1;
-	}
-
-	hdr = (struct ip2gid_hdr *)obj->data;
-	ip2gid_log(IP2GID_LOG_INFO,
-		   "Checking new Msg (msg_id = %u)\n",
-		   ntohl(hdr->msg_id));
-	num_tlvs = ntohs(hdr->num_tlvs);
-	if (!num_tlvs) {
-		ip2gid_log(IP2GID_LOG_INFO,
-			   "\tNo TLVS in Msg\n");
-		return -1;
-	}
-
-	ip2gid_log(IP2GID_LOG_INFO, "\tNum tlvs:%u\n", num_tlvs);
-	if (hdr->version > IP2GID_CURRENT_VERSION) {
-		ip2gid_log(IP2GID_LOG_INFO,
-			   "\tMsg version isn't supported\n");
-		return -1;
-	}
-
-	max_len -= sizeof(struct ip2gid_hdr);
-
-	tlv = (struct ip2gid_tlv_hdr *)hdr->tlvs;
-	while (num_tlvs && max_len >= 0) {
-		uint16_t tlv_len = ntohs(tlv->len);
-
-		max_len -= tlv_len;
-		switch (ntohs(tlv->type)) {
-		case IP2GID_REQ_NONE:
-			if (tlv_len != sizeof(struct ip2gid_tlv_hdr)) {
-				ip2gid_log(IP2GID_LOG_INFO,
-					   "\tTLV: None: invalid length (len= %u)\n",
-					   tlv_len);
-				return -1;
-			}
-			break;
-		case IP2GID_REQ_IPV4:
-			if (tlv_len != sizeof(struct ip2gid_req_ipv4)) {
-				ip2gid_log(IP2GID_LOG_INFO,
-					   "\tTLV: IPV4: invalid length (len = %u)\n",
-					   tlv_len);
-				return -1;
-			}
-			break;
-		case IP2GID_RESP_GID:
-			if (tlv_len != sizeof(struct ip2gid_resp_gid)) {
-				ip2gid_log(IP2GID_LOG_INFO,
-					   "\tTLV: GID: invalid length (len = %u)\n",
-					   tlv_len);
-				return -1;
-			}
-			break;
-		default:
-			ip2gid_log(IP2GID_LOG_INFO,
-				   "\tTLV: Unknown type\n");
-			return -1;
-		}
-		tlv = (struct ip2gid_tlv_hdr *)((char *)tlv + tlv_len);
-		num_tlvs--;
-	}
-	if (num_tlvs == 0 && max_len >= 0) {
-		ip2gid_log(IP2GID_LOG_INFO,
-			   "\tMsg passed basic checks\n");
-		return 0;
-	}
-
-	return -1;
-}
-
 static int server_find_gid(struct ip2gid_req_ipv4 *req,
 			   struct ip2gid_hdr *resp_hdr,
 			   struct ip2gid_resp_gid *resp,
